Downloads/real-estate-backend/requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
psycopg2-binary>=2.9
djangorestframework>=3.14.0
python-dotenv
pandas
supabase
